Refactoring to use custom hook for project data
