+++ src/components/organisms/projects/JoinProject.tsx
@@ -2,8 +2,8 @@ import { FC, useEffect, useRef, useState } from 'react';
 import { ProjectService } from '../../../services/projects.service';
 import { supabase } from '../../../utils/supabaseClient';
 import { PrimaryButton } from '../../atoms/button/PrimaryButton';
-import { t } from '../../../utils/commonTools';
 import { debug } from '../../../utils/commonTools';
+import { useT } from '@transifex/react';
 
 type Props = {
   joined?: boolean;
@@ -13,6 +13,8 @@ type Props = {
 const MJoinProject: FC<Props> = ({ joined = false, project_id }) => {
   const [isJoined, setIsJoined] = useState(joined);
   const requesting = useRef(false);
+  const t = useT();
+
   useEffect(() => {
     setIsJoined(joined);
   }, [joined]);
@@ -20,7 +22,7 @@ const MJoinProject: FC<Props> = ({ joined = false, project_id }) => {
   const handleClick = async () => {
     const user = supabase.auth.user();
     if (!user) {
-      alert('ログインしてね。');
+      alert(t('you need to login first'));
       return;
     }
     if (isJoined === null || requesting.current) return;
