+++ src/components/organisms/projects/Members.tsx
@@ -1,18 +1,20 @@
 import { Box, Heading } from '@chakra-ui/react';
 import { FC } from 'react';
-import { definitions } from '../../../../types/supabase';
+import { definitions } from '../../../types/supabase';
 
 type Props = {
-  members: definitions['profiles'][];
+  members: definitions['profiles'][] | undefined;
 };
 
 const MMembers: FC<Props> = ({ members }) => {
   return (
     <Box>
       <Heading as='h2'>OurMembers</Heading>
-      {members.map((member, index) => {
-        return <div key={index}>{member.username}</div>;
-      })}
+      {members
+        ? members.map((member, index) => {
+            return <div key={index}>{member.username}</div>;
+          })
+        : ''}
     </Box>
   );
 };
