+++ src/pages/projects/[id].tsx
@@ -1,7 +1,11 @@
 import { Box, Flex } from '@chakra-ui/react';
 import { useRouter } from 'next/router';
-import SProject from '../../components/SProject';
-export default function Project() {
+import SProject from '../../components/organisms/projects/SProject';
+
+/**
+ * project component
+ */
+export default function ProjectPage() {
   const router = useRouter();
   const { id } = router.query;
   return (
