+++ src/pages/projects/new.tsx
@@ -1,12 +1,13 @@
 import { supabase } from '../../utils/supabaseClient';
 import { Flex } from '@chakra-ui/react';
-import ProjectForm from '../../components/ProjectForm';
+import ProjectForm from '../../components/organisms/projects/ProjectForm';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 export default function CreateProject() {
   const router = useRouter();
 
   useEffect(() => {
+    // if user is not logged in, redirect to the index
     if (!supabase.auth.user()) {
       router.replace('/');
     }
